Getting ready for compression
Everything seems to not be going by that bad, the only thing I know that I am going to have to account for is the possibly undefined actions of the rotation functions, as the with using a 32 bit number in a 64 bit wrapper, there are some challenges that I am going to have to face, and I think this is one of them. I think I have to modify the shifting a little in order to make sure none of the garbage that is in the most significant bits will be mucking up the calculation. I really don't want to need to set the block to zero and then operate as that would require more computation time. We will see. Working on 32bit_with64bit_wrapper.c git add .!