Working!
I still need to work on the getready_readoutput fully ironed out with the edge cases (already have the algo from past attempt), but I still need to implement it. I also changed the H_access array, as I realized that the older way I was going about doing it, required me to change the pointer of the 'head', but What I ended up doing is more akin to just a RIGHT ROTATE. I still have to check the speed difference, but I am quite curious to see what the veredict is!