Finished, looking to optimiize
I am finished and the program is working correctly, as it will return the correct hash even with 'very large' data sets like a windows 10 iso. Though I think that there is more optimiztion to squeez out of the program. I think the part that could imporve the most is the reading. As of right now, an idea is to try to read a page size out of the file at a time instead of having to read 4 bytes endlessly at a time. I think readying 4 bytes at a time is calling the fread function unecesarily.